Move dedicated function into component
